Renamed lifecycleweather to usdaplantindex; begin StartupActivity
