+++ Code/app/src/main/AndroidManifest.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.android.lifecycleweather">
+    package="com.example.android.usdaplantindex">
+
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
@@ -9,9 +11,13 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.android.usdaplantindex.StartupActivity"
+            android:label="@string/title_activity_startup">
+
+        </activity>
+        <activity
+            android:name="com.example.android.usdaplantindex.MainActivity"
             android:launchMode="singleTop">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -19,17 +25,13 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
         <activity
-            android:name=".PlantItemDetailActivity"
-            android:parentActivityName=".MainActivity">
+            android:name="com.example.android.usdaplantindex.PlantItemDetailActivity"
+            android:parentActivityName="com.example.android.usdaplantindex.MainActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value="com.example.android.usdaplantindex.MainActivity" />
         </activity>
-
     </application>
-
-    <uses-permission android:name="android.permission.INTERNET" />
 
 </manifest>
