+++ Code/app/src/main/java/com/example/android/usdaplantindex/PlantAdapter.java
@@ -1,4 +1,4 @@
-package com.example.android.lifecycleweather;
+package com.example.android.usdaplantindex;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -7,7 +7,8 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.example.android.lifecycleweather.utils.USDAUtils;
+import com.example.android.usdaplantindex.R;
+import com.example.android.usdaplantindex.utils.USDAUtils;
 
 import java.util.ArrayList;
