+++ Code/app/src/main/java/com/example/android/usdaplantindex/utils/NetworkUtils.java
@@ -1,4 +1,4 @@
-package com.example.android.lifecycleweather.utils;
+package com.example.android.usdaplantindex.utils;
 
 import java.io.IOException;
